+++ lib/src/features/super_table/super_table.dart
@@ -25,6 +25,7 @@ export 'domain/entities/super_row.dart';
 export 'domain/entities/super_style.dart';
 export 'domain/entities/super_change.dart';
 export 'domain/entities/super_filter.dart';
+export 'domain/entities/super_group.dart';
 export 'domain/entities/super_table_state.dart';
 export 'domain/usecases/super_column_logic.dart';
 export 'presentation/controllers/super_table_controller.dart';